feat: code clean with directory
