+++ src/pages/CardCreator/InputComponents/CardOwnerInput/index.ts
@@ -1,0 +1,1 @@
+export * from './CardOwnerInput';
