+++ src/pages/CardCreator/InputComponents/ExpireDatesInputList/ExpireDateInput.stories.tsx
@@ -6,25 +6,24 @@ import '@/styles/input.css';
 import React, { useState } from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { cardNumbersInit } from '../CardCreatorInits';
-import { CardNumberInputPure } from './CardNumberInput';
-import { createCardStateSetter } from '../utils';
+import { expireDatesInit } from '../../CardCreatorInits';
+import { ExpireDatesInputListPure } from './ExpireDatesInputList';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'CardCreator/CardNumberInput',
-  component: CardNumberInputPure,
+  title: 'CardCreator/ExpireDateInput',
+  component: ExpireDatesInputListPure,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     cardNumbers: { control: false },
     createCardNumberSetter: { control: false },
   },
-} as ComponentMeta<typeof CardNumberInputPure>;
+} as ComponentMeta<typeof ExpireDatesInputListPure>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof CardNumberInputPure> = () => {
-  const [cardNumber, setCardNumber] = useState(cardNumbersInit);
-  return <CardNumberInputPure cardNumbers={cardNumber} createCardNumberSetter={createCardStateSetter(setCardNumber)} />;
+const Template: ComponentStory<typeof ExpireDatesInputListPure> = () => {
+  const [expireDates, setExpireDates] = useState(expireDatesInit);
+  return <ExpireDatesInputListPure expireDates={expireDates} />;
 };
 
 export const Primary = Template.bind({});
