+++ src/pages/CardCreator/InputComponents/PasswordsInputList/PasswordInput/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordInput';
