+++ src/pages/CardCreator/InputComponents/PasswordsInputList/PasswordsInputList.stories.tsx
@@ -6,25 +6,24 @@ import '@/styles/input.css';
 import React, { useState } from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { passwordsInit } from '../CardCreatorInits';
-import { PasswordInputPure } from './PasswordInput';
-import { createCardStateSetter } from '../utils';
+import { passwordsInit } from '../../CardCreatorInits';
+import { PasswordsInputListPure } from './PasswordsInputList';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
   title: 'CardCreator/PasswordInput',
-  component: PasswordInputPure,
+  component: PasswordsInputListPure,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     cardNumbers: { control: false },
     createCardNumberSetter: { control: false },
   },
-} as ComponentMeta<typeof PasswordInputPure>;
+} as ComponentMeta<typeof PasswordsInputListPure>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof PasswordInputPure> = () => {
+const Template: ComponentStory<typeof PasswordsInputListPure> = () => {
   const [passwords, setPasswords] = useState(passwordsInit);
-  return <PasswordInputPure passwords={passwords} createPasswordSetter={createCardStateSetter(setPasswords)} />;
+  return <PasswordsInputListPure passwords={passwords} />;
 };
 
 export const Primary = Template.bind({});
