+++ src/pages/CardCreator/InputComponents/PasswordsInputList/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordsInputList';
