--- src/pages/CardCreator/InputComponents/index.ts
@@ -1,5 +1,0 @@
-export * from './CardNumberInput';
-export * from './CardOwnerInput';
-export * from './ExpireDateInput';
-export * from './PasswordInput';
-export * from './SecurityCodeInput';
